simpleify base framework
