--- base-framework/agents/goodbye.py
@@ -1,2 +1,0 @@
-def agent_main():
-    return "Goodbye from the agent!"
