+++ base-framework/app/main.py
@@ -5,28 +5,15 @@ from typing import Optional, List, Dict, Any
 import os
 from agents.dspy_integration import load_agent, run_agent
 from agents.classifier import register_routes as register_classifier_routes
-from agents.summarizer import register_routes as register_summarizer_routes
-from agents.textrank_summarizer import register_routes as register_textrank_summarizer_routes  # NEW
-from agents.echo import register_routes as register_echo_routes              # NEW
-from agents.time import register_routes as register_time_routes              # NEW
-from agents.joke import register_routes as register_joke_routes              # NEW
 from agents.quote import register_routes as register_quote_routes            # NEW
-from agents.math import register_routes as register_math_routes              # NEW
 
-app = FastAPI(title="FastAPI Agent System")
+app = FastAPI(title="FastAPI Agent System - Basic Framework - github.com/bar181")
 
 # --- Agent Information ---
 AGENTS_INFO: List[Dict[str, str]] = [
-    {"name": "hello_world", "description": "Returns a simple hello world message."},
-    {"name": "goodbye", "description": "Returns a goodbye message."},
-    {"name": "echo", "description": "Returns a simple echo message."},
-    {"name": "time", "description": "Returns the current time in ISO 8601 format."},
-    {"name": "joke", "description": "Returns a random programming joke."},
+    {"name": "hello_world", "description": "Returns a simple hello world message."},   
     {"name": "quote", "description": "Returns an inspirational quote."},
-    {"name": "math", "description": "Evaluates a math expression after verifying a token."},
     {"name": "classifier", "description": "Classifies input text using rule-based logic."},
-    {"name": "summarizer", "description": "Summarizes a block of text (truncation)."},
-    {"name": "textrank_summarizer", "description": "Summarizes text using TextRank algorithm."},
 ]
 
 @app.get("/agents", tags=["All Agents"])
@@ -35,35 +22,15 @@ async def list_all_agents() -> Dict[str, List[Dict[str, str]]]:
 
 # --- Agent Router ---
 agent_router = APIRouter(prefix="/agent")
-register_classifier_routes(agent_router)
-register_summarizer_routes(agent_router)
-register_textrank_summarizer_routes(agent_router)  # NEW: Register the TextRank routes
-register_echo_routes(agent_router)                 # NEW: Register Echo Agent routes
-register_time_routes(agent_router)                 # NEW: Register Time Agent routes
-register_joke_routes(agent_router)                 # NEW: Register Joke Agent routes
-register_quote_routes(agent_router)                # NEW: Register Quote Agent routes
-register_math_routes(agent_router)                 # NEW: Register Math Agent routes
+register_classifier_routes(agent_router)           # DSPY: Use case for dspy 
+
+register_quote_routes(agent_router)                # Simple: Basic agent
 app.include_router(agent_router)
 
 
 # --- Other Routes (hello_world, goodbye, generic) ---
-# Keep hello_world and goodbye as before.  The generic route is also important.
 
-@app.get("/agent/hello_world", tags=["Simple Agents"])
-async def hello_world_agent():
-    agent_file = os.path.join("agents", "hello_world.py")
-    agent_module = load_agent(agent_file)
-    output = run_agent(agent_module)
-    return {"agent": "hello_world", "result": output}
-
-@app.get("/agent/goodbye")
-async def goodbye_agent():
-    agent_file = os.path.join("agents", "goodbye.py")
-    agent_module = load_agent(agent_file)
-    output = run_agent(agent_module)
-    return {"agent": "goodbye", "result": output}
-
-
+# dymanic agent generation using dspy
 @app.get("/agent/{agent_name}", tags=["Dynamic Agents"]) 
 async def execute_agent(agent_name: str, request: Request):
     agent_file = os.path.join("agents", f"{agent_name}.py")
@@ -100,7 +67,7 @@ async def get_favicon():
 
 @app.get("/")
 async def read_root():
-    return {"message": "Welcome to the Agent System!"}
+    return {"message": "Welcome to the Agent Base Framework! (https://github.com/bar181/fastapi-agents)"}
 
 @app.get("/health")
 async def health_check():
