+++ base-framework/plans/plans.md
@@ -1,3 +1,50 @@
-## step by step process for each phase of your plan
+# Guide to the /plans Section
 
-use naming convention <step integer>-<phase description>.md             e.g. 1-framework-setup.md
+## Purpose of the /plans Section
+
+The `/plans` directory contains structured phase-based planning documents that outline the development process for a project. Each document follows a standardized naming convention:
+
+```
+<step integer>-<phase description>.md
+```
+
+Example:
+```
+1-framework-setup.md
+```
+
+Each plan document serves as a blueprint for a specific phase of development, ensuring a structured and well-documented approach to project execution.
+
+## How /plans Fit into the Development Process
+
+Our development process follows a rigorous **Documentation First** methodology, structured as follows:
+
+1. **Brainstorm and Design:**  
+   Initial ideas and designs are discussed and documented to establish project goals.
+
+2. **Required Documentation:**  
+   All technical and implementation documents are created and stored in the `/docs` directory.
+
+3. **Phase Plans:**  
+   Detailed execution plans for each phase of development are stored in the `/plans` directory.
+
+4. **Step-by-Step Execution:**  
+   Development is carried out according to the structured plans, including comprehensive testing.
+
+5. **Tracking and Logging:**  
+   Every step is tracked and logged in the `/logs` directory for full transparency and reproducibility.
+
+6. **Testing and Updates:**  
+   Proper testing is implemented, and documentation is continuously updated to reflect the current state of the project.
+
+## Structure of a Plan Document
+
+Each plan document within the `/plans` directory should have full details for each individual step of your process:
+
+---
+
+
+
+## Summary
+
+The `/plans` directory is a critical component of the structured development process. By adhering to the standardized format and following the Documentation First methodology, development teams ensure consistency, traceability, and efficiency in project execution. Each plan document provides clear, actionable steps to guide implementation while maintaining detailed documentation and logging for future reference.
