+++ base-framework/tests/test_main.py
@@ -3,21 +3,21 @@ from app.main import app
 
 client = TestClient(app)
 
-def test_read_root():
-    """Test the root endpoint returns welcome message"""
-    response = client.get("/")
-    assert response.status_code == 200
-    assert response.json() == {"message": "Welcome to the Agent System!"}
-
 def test_health_check():
     """Test the health check endpoint"""
     response = client.get("/health")
     assert response.status_code == 200
     assert response.json() == {"status": "ok", "message": "Healthy"}
 
-def test_favicon():
-    """Test the favicon endpoint returns SVG"""
-    response = client.get("/favicon.ico")
-    assert response.status_code == 200
-    assert response.headers["content-type"] == "image/svg+xml"
-    assert b"svg" in response.content
+def test_nonexistent_agent():
+    """Test error handling for non-existent agent"""
+    response = client.get("/agent/nonexistent")
+    assert response.status_code == 404
+    assert "Agent not found" in response.json()["detail"]
+
+def test_invalid_agent():
+    """Test error handling for invalid agent file"""
+    # This would require creating a temporary invalid agent file
+    # For now, we'll just verify the 404 response
+    response = client.get("/agent/invalid")
+    assert response.status_code == 404
