+++ dspy/.env.sample
@@ -1,2 +1,1 @@
 # List of .env secrets required for this module
-# none
